Fix "'dict object' has no attribute 'mariadb'" with Ansible 2.2
`checkmk_agent__plugin_mysql_password` references the Ansible local facts
`ansible_local.mariadb` which is undefined if the role is skipped via
condition in a playbook. Since Ansible 2.2 this results in an error.
This fix will make sure, that the local fact is only used when defined.

Fixes debops-contrib/ansible-checkmk_agent#22